// src/types/bank-account.ts
export interface BankAccount {
  id: string;
  // holder: string;
  name:string;
  mask: string;
  holder_category:string;
  type: string;
  subtype: string;
  current_balance:string;
  limit_balance:string;
  iso_currency_code: string;
  official_name: string;
  createdAt: string;
  updatedAt: string;
}